user-service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
requests>=2.26.0
pydantic>=1.8.0
opentelemetry-api>=1.11.0
opentelemetry-sdk>=1.11.0
opentelemetry-instrumentation-fastapi>=0.30b0
opentelemetry-instrumentation-sqlalchemy>=0.30b0
opentelemetry-exporter-otlp>=1.11.0
prometheus-client>=0.14.0
circuitbreaker>=1.3.0